% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkConvergence.R
\name{checkConvergence}
\alias{checkConvergence}
\title{Convergence test.}
\usage{
checkConvergence(mcmc.chain, out)
}
\arguments{
\item{mcmc.chain}{The MCMC chain. Output of the function to read the chain. It works both with any number of matrices fitted to the tree.}

\item{out}{The MCMC result output. This is a list with several details about the MCMC run.}
}
\value{
A matrix with colums varying dependent of the number of matrices fitted to the tree. The matrix
   has two lines, each is a component of the Heidel diagnostic. TRUE is when the convergence test passed and
   FALSE is when it falied.
}
\description{
Make convergence test for the MCMC chain.
}
\details{
This function makes convergence test for the MCMC chains. It uses the Heidel test from the package 'coda'.
   A better test of convergence is the Gelman's R, but this test requires more than one MCMC chain,
   recommended to start for distinct start values of the chain for better accessing the convergence.
   This other test will be implemented soon.
To calculate the convergence this function will check if each value of each cell of the matrices have
   converged. It is hard to find a better way to check for convergence. This is the same approach used by
   MCMCglmm and the function to test the significance of the difference between two posterior distributions
   of matrices. This approach is most likely more conservative than would be looking to the matrix as
   a whole.
}

