% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkHPD.R
\name{checkHPD}
\alias{checkHPD}
\title{Check if matrix is within HPD}
\usage{
checkHPD(mat, qq, dd)
}
\arguments{
\item{mat}{matrix or dataframe.}

\item{qq}{list. Elements are vectors of length 2 with range of the hpd.}

\item{dd}{numeric. The dimension of the matrix.}
}
\value{
boolean. Return TRUE when all elements of 'mat' is within the interval of 'qq'.
}
\description{
Check if all elements are in the HPD (Highest Posterior Density) interval. Internal function to be used by 'make.grid.plot'.
}
\details{
This funtion is not optimized to be called by the user.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}
\seealso{
\code{\link{make.grid.plot}}
}
\keyword{utilities}

