% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeESS.R
\name{computeESS}
\alias{computeESS}
\title{Compute the ESS for the MCMC samples.}
\usage{
computeESS(mcmc, p)
}
\arguments{
\item{mcmc}{Posterior distribution object. Same as output from 'readMCMC' function.}

\item{p}{Number of evolutionary rate matrix regimes fitted to the phylogenetic tree.}
}
\value{
A list object with the Effective Sample Size for the root values, evolutionary rates, and evolutionary correlations among the traits.
}
\description{
Function uses 'coda' function 'effectiveSize' to compute the ESS for each of the parameters of the model separatelly.
}
\author{
Daniel Caetano and Luke Harmon
}
