% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continueMCMC.R
\name{continueMCMC}
\alias{continueMCMC}
\title{Make the thing!}
\usage{
continueMCMC(handle, add.gen = NULL, save.handle = TRUE)
}
\arguments{
\item{handle}{}

\item{add.gen}{}

\item{save.handle}{}
}
\value{
The thing!
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
daniel
}

