% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.mcmc.R
\name{dic.mcmc}
\alias{dic.mcmc}
\title{Deviance Information Criteria}
\usage{
dic.mcmc(out, post)
}
\arguments{
\item{out}{The object with the MCMC info. The return from the MCMC functions.}

\item{post}{The posterior chain read by the 'read.multi.R.iwish' or 'read.single.R.iwish'.}
}
\value{
numeric. Return the DIC value.
}
\description{
Deviance information criterion DIC for mcmcsamples.
}
\details{
Function works with the samples from the R matrix MCMC estimate functions.
Future implementations will have the DIC calculated alongside the MCMC chain.\cr
\cr
Issue: DIC is prone to missbehave and reject the model with two or more matrixes even when there is
     clearly a better estimate of the parameters. We recommend to use DIC with caution. One can instead
     interpret the posterior distribution of matrices and check whether the matrices significantly
     overlap.
}
\seealso{
\code{\link{multi.R.iwish.mcmc.R,single.R.iwish.mcmc.R}}
}

