% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinvchisq.R
\name{dinvchisq}
\alias{dinvchisq}
\title{Log density of the scaled inverse chi-squared}
\usage{
dinvchisq(x, v.chi, scale = 1/v.chi)
}
\arguments{
\item{x}{vector. Vector of variances.}

\item{v.chi}{numeric. Degrees of freedom.}

\item{scale}{numeric. Scale for the distribution.}
}
\value{
numeric. Log density of the scaled inverse chi-squared.
}
\description{
Log density of the scaled inverse chi-squared distribution
}
\details{
Calculates the density of the scaled inverse chi-squared distribution in log space. \cr
\cr
The parameter 'scale' has a default value of 1/v.chi.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

