% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective.dep.R
\name{effective.dep}
\alias{effective.dep}
\title{Effective dependence.}
\usage{
effective.dep(mat, burn = 0.5, thin = 1)
}
\arguments{
\item{mat}{list. list of covariance matrices.}

\item{burn}{numeric. proportion of the burnin.}

\item{thin}{numeric. the thinning to be applied in the posterior distribution. Same of the argument 'by' of the function 'seq'.}
}
\value{
numeric. Return a vector with the effective dependence index of the matrices.
}
\description{
Calculates the effective dependence of a list of matrices.
}
\details{
This funtion is not optimized to be called by the user.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

