% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effective.var.R
\name{effective.var}
\alias{effective.var}
\title{Effective variance.}
\usage{
effective.var(mat, burn = 0.5, thin = 1)
}
\arguments{
\item{mat}{list. list of covariance matrices.}

\item{burn}{numeric. proportion of the burnin.}

\item{thin}{numeric. the thinning to be applied in the posterior distribution. Same of the argument 'by' of the function 'seq'.}
}
\value{
numeric. Return a vector with the effective variance index of the matrices.
}
\description{
Calculate the effective variance of a list of matrices.
}
\details{
This funtion is not optimized to be called by the user.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

