% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.chart.R
\name{ellipse.chart}
\alias{ellipse.chart}
\title{Ellipse plot of covariance matrices.}
\usage{
ellipse.chart(mat, mu, burn = 0.5, thin = 1, sample.line = 50,
  n.points = 200, center.mu = TRUE, true.R = NULL, file = NULL)
}
\arguments{
\item{mat}{list. List of covariance matrices.}

\item{mu}{matrix. Matrix of mean values to use as centroid for each ellipse.}

\item{burn}{numeric. Proportion of burnin.}

\item{thin}{numeric. The thinning of the posterior distribution. Same as the argument 'by' of the function 'seq'.}

\item{sample.line}{numeric. Number of ellipse lines to be plotted in each plot.}

\item{n.points}{numeric. Number of points to be used to calculate the shape of the ellipses.}

\item{center.mu}{logical. Whether the center of the ellipses should be guven by the argument 'mu'. Otherwise all ellipses are centered in the '0,0' coordinate.}

\item{true.R}{matrix. A single evolutinary to be plotted as a red line. Usually the true value of a set of simulations.}

\item{file}{string. When provided the functions creates a '.jpeg' figure instead of plotting in the standard device.}
}
\value{
Plot in standard device or a '.jpeg' file. See argument 'file'.
}
\description{
Make plots for all elements in a list (posterior distribution) of evolutionary covariance matrices (R).
}
\details{
The function make a grid of plots with ellipses representing the distribution of each variance and covariance element
    of a list of evolutionary covariance matrices (R). Disposition of the plot is calculated to prefer spread over the
    columns while minimizing number of rows.\cr
\cr
When a matrix for 'true.R' is provided, function use this matrix to create red lines and superpose them above the distribution
    of ellipses.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

