% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTimeMCMC.R
\name{estimateTimeMCMC}
\alias{estimateTimeMCMC}
\title{All the things!}
\usage{
estimateTimeMCMC(data, phy, gen, eval.times = 5, singlerate = FALSE)
}
\arguments{
\item{data}{}

\item{phy}{}

\item{gen}{}

\item{eval.times}{}

\item{singlerate}{}
}
\value{
The thing.
}
\description{
Do the thing!
}
\details{
Really does it!
}
\author{
daniel
}

