% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vcv.ellipse.all.matrix.R
\name{get.vcv.ellipse.all.matrix}
\alias{get.vcv.ellipse.all.matrix}
\title{Calculate all ellipses.}
\usage{
get.vcv.ellipse.all.matrix(mat, center = c(0, 0), traits, n.points = 200)
}
\arguments{
\item{mat}{list. List of covariance matrices.}

\item{center}{numeric. Coordinates for the center.}

\item{traits}{numeric (?). The traits (?).}

\item{n.points}{numeric. The number of points used to approximate the shape of the ellipse.}
}
\value{
List of coordinates to plot the ellipses.
}
\description{
Calculates points for ellipses.
}
\details{
This funtion is not optimized to be called by the user. Function calculates the points to draw ellipses from a list of covariance matrices.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

