% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.vcv.ellipse.matrix.R
\name{get.vcv.ellipse.matrix}
\alias{get.vcv.ellipse.matrix}
\title{Calculate sample of ellipses.}
\usage{
get.vcv.ellipse.matrix(mat, center = c(0, 0), traits, burn = 0.5,
  thin = 1, sample.line = 50, n.points = 200)
}
\arguments{
\item{mat}{list. List of covariance matrices.}

\item{center}{numeric. Coordinates for the center.}

\item{traits}{numeric (?). The traits (?).}

\item{burn}{numeric. The proportion of the burnin.}

\item{thin}{numeric. The thinnig to be applied to the posterior distribution. Same as the 'by' argument of the 'seq' function.}

\item{sample.line}{numeric. Sample size to be used to calculate the ellipses.}

\item{n.points}{numeric. The number of points used to approximate the shape of the ellipse.}
}
\value{
List of coordinates to plot the ellipses.
}
\description{
Calculates points for ellipses.
}
\details{
This funtion is not optimized to be called by the user. Function calculates the points to draw ellipses from a list of covariance matrices.\cr
\cr
If 'mat' is only a matrix then only the coordinates for the plot is returned. Otherwise, a list with the coordinates and limits calculated for
the series of ellipses is returned.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

