% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hastingsDensity.R
\name{hastingsDensity}
\alias{hastingsDensity}
\title{log Hastings ratio}
\usage{
hastingsDensity(curr.vcv, prop.vcv, p, v)
}
\arguments{
\item{curr.vcv}{matrix. The current evolutionary covariance matrix.}

\item{prop.vcv}{matrix. The proposed evolutionary covariance matrix.}

\item{p}{numeric. Number of traits in the analysis.}

\item{v}{numeric. Variance parameter of the inverse wishart distrubution.}
}
\value{
numeric. The log Hastings ratio for the inverse wishart based proposal distribution.
}
\description{
Calculates the log Hastings ratio.
}
\details{
Calculates the log Hastings ratio by dividing p(curr|prop)/p(prop|curr) densities. Internal function of the MCMC.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

