% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.rate.R
\name{hist.rate}
\alias{hist.rate}
\title{Histogram of rates.}
\usage{
\method{hist}{rate}(x, main = NULL)
}
\arguments{
\item{x}{list. Posterior distribution of the evolutionary covariance matrix (R).}

\item{main}{string. Title for the plot.}
}
\value{
Plot.
}
\description{
Plot histogram of rates.
}
\details{
Make a custom histogram to plot the posterior distribution of the rates.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

