% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunction.R
\name{likelihoodFunction}
\alias{likelihoodFunction}
\title{Likelihood function for the multivariate Brownian motion model.}
\usage{
likelihoodFunction(data, phy, root, R)
}
\arguments{
\item{data}{a matrix with the data. Species names need to be provided as rownames (rownames(data) == phy$tip.label).}

\item{phy}{a phylogeny of the class "simmap" with the mapped regimes. The number of evolutionary rate matrices fitted to the phylogeny need to be equal to the number of regimes in phy.}

\item{root}{a numeric vector with the root value (phylogenetic mean).}

\item{R}{a matrix or a list of matrices. If 'R' is a matrix then the likelihood for a single regime is calculated. If 'R' is a list of matrices, then each matrix will be fitted to a regime in 'phy'. The R matrices will be fitted to the regimes in the same order as the columns of 'phy$mapped.edge'.}
}
\value{
The log likelihood for the multivariate Brownian motion model.
}
\description{
Function calculates the log-likelihood for the multivariate Brownian motion model given a phylogenetic tree with rate regimes mapped to it, a data matrix with the trait value for the species, a vector with the phylogenetic means (root values) and a matrix (or list of matrices) with the evolutionary rate matrix for each regime.
}
\details{
If more than one evolutionary rate matrix is used, then the function calculates the likelihood using the multirates prunning algorithm. Otherwise the function uses the three point algorithm to make calculations for the single regime case.
}
\author{
daniel
}

