% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.diwish.R
\name{log.diwish}
\alias{log.diwish}
\title{Log density for the inverse Wishart.}
\usage{
\method{log}{diwish}(W, v, S)
}
\arguments{
\item{W}{matrix. The covariance matrix.}

\item{v}{numeric. The degrees of freedom parameter.}

\item{S}{matrix. The standard matrix for the distribution.}
}
\value{
numeric. The log density.
}
\description{
Calculates the log density of the inverse Wishart distribution.
}
\details{
Modified from the 'diwish' function of the package 'MCMCpack'.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

