% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logDensityMvNorm.R
\name{logDensityMvNorm}
\alias{logDensityMvNorm}
\title{Log density of the multivariate normal.}
\usage{
logDensityMvNorm(x, mu, sigma = diag(length(x)))
}
\arguments{
\item{x}{numeric. Vector of mean values.}

\item{mu}{numeric. Vector of the distribution means.}

\item{sigma}{matrix. Variance covariance matrix parameter.}
}
\value{
numeric. Log density.
}
\description{
Log density for the multivariate normal distrubution.
}
\details{
The default for 'sigma' is the identity matrix.
}
\author{
Daniel S. Caetano, \email{caetanods1@gmail.com}
}

