% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikMultRegime.R
\name{logLikMultRegime}
\alias{logLikMultRegime}
\title{Log likelihood for multiple matrices.}
\usage{
logLikMultRegime(data, chain, root, R)
}
\arguments{
\item{data}{cache data.}

\item{chain}{cache chain}

\item{root}{phylogenetic mean}

\item{R}{list. List of the R matrices.}
}
\value{
numeric. The log likelhood.
}
\description{
Log likelihood function for two or more evolutionary rate matrices.
}
\details{
Internal function to be used by the MCMC. This function uses the 'rpf' method to calculate the approximation of the inverse of matrices as impplemented in mvMORPH.
}
\references{
Revell, L. J., and D. C. Collar. 2009. Phylogenetic Analysis of the Evolutionary Correlation Using Likelihood. Evolution 63:1090–1100.
}

