% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLikSingleRegime.R
\name{logLikSingleRegime}
\alias{logLikSingleRegime}
\title{Log likelihood for one evolutionary rate matrix fitted to the data and tree.}
\usage{
logLikSingleRegime(data, chain, phy, root, R)
}
\arguments{
\item{data}{the data cache.}

\item{chain}{the chain cache.}

\item{root}{numeric. The phylogenetic mean.}

\item{R}{matrix. The covariance matrix fitted to the tree and data.}
}
\value{
numeric. The log likelhood.
}
\description{
Log likelihood function for one evolutionary rate matrix.
}
\details{
This function applies a math shortcut to avoid inverting the kronecker product of matrices. The trick is based on an equality described in Ho and Ané (2014). A similar equality also make the calculation of the determinant of the kronecker product faster. Thus, one need to provide the inverse of the phylogenetic covariance matrix as the argument 'C.prime' and the determinant of this same matrix as the argument 'det.C'.
}
\references{
Revell, L. J., and D. C. Collar. 2009. Phylogenetic Analysis of the Evolutionary Correlation Using Likelihood. Evolution 63:1090–1100.

Ho, L. si T., and C. Ané. 2014. A Linear-Time Algorithm for Gaussian and Non-Gaussian Trait Evolution Models. Syst Biol 63:397–408.
}

