% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.grid.plot.R
\name{make.grid.plot}
\alias{make.grid.plot}
\title{Plot posterior distribution of rate matrices.}
\usage{
make.grid.plot(mat1, mat2 = NULL, mle1 = NULL, mle2 = NULL,
  colDiag1 = "purple", colDiag2 = "green", alphaOff1 = 0.5,
  alphaOff2 = 0.5, leg = NULL, l.cex = 0.7, hpd = 100,
  show.zero = FALSE, set.xlim = NULL, n.lines = 50)
}
\arguments{
\item{mat1}{list. The posterior distribution of the R matrix.}

\item{mat2}{list. The posterior distribution of the R matrix to be plotted with 'mat1'.}

\item{mle1}{matrix. Used as a point estimate. Will be plotted as a vertical line or a single ellipse. Color is 'red'.}

\item{mle2}{matrix. Same as 'mle2'. Can be used as the point estimate for 'mat2'.}

\item{colDiag1}{string. Color for the diagonal of the grid plot and the lines of the lower-tri ellipses.}

\item{colDiag2}{string. Same as 'colDiag1'. Can be used to set color of the 'mat2' plots.}

\item{alphaOff1}{numeric. Transparency of the off-diagonal plots. Value between 0 and 1.}

\item{alphaOff2}{numeric. Same as 'alphaOff1'. Set the transparency of the plots relative of 'mat2'.}

\item{leg}{string. Legend for the traits. Vector need to have length equal to the dimension of the R matrix.}

\item{l.cex}{numeric. 'cex' parameter for 'leg'. See 'help(par)' for more information on 'cex'.}

\item{hpd}{numeric. Set the proportion of the highest posterior density (HPD) to be highlighted in the plot. If set .95 the distributions and the ellipses within the 95% HPD will be highlighted in the plot. Default value does not show highlighted region.}

\item{show.zero}{logical. Whether to plot a 'blue' vertical line at 0.}

\item{set.xlim}{numeric. Two elements vector to set the 'xlim' [see 'help(hist)'] for the density plots manually. If 'NULL', the default value, then the limits are calculated from the data.}

\item{n.lines}{numeric. Number of lines to be displayed in the ellipsed plots.}
}
\value{
Plot a grid of charts.
}
\description{
Make a grid plot with the posterior distribution of the evolutionary rate matrix (R). Function works with one or two matrices fitted to the same tree and will superpose the results.
}
\details{
The upper-tri plots are the posterior density for the off-diagonal elements of the
    R evolutionary matrix. The diagonal plots are the posterior elements of the evolutionary
    rates for each of the traits. The lower-tri plots are ellipse plots for the rates of correlated
    evolution.\cr
\cr
Two posterior distributions of matrices can be plotted in the same grid if a distribution for
    'mat2' is also provided (optional). Otherwise only 'mat1' is plotted.\cr
\cr
If a second distribution of matrices is provided ('mat2') the function will check whether the number
     of matrices in each list is the same. If not, the first elements of the lengthier list will be pruned
     to match the length of the shorter list.\cr
\cr
Function use 'par()' to set graphical parameters. It may take some time to create all the plots if
     the number of traits is large. Thinning the posterior distribution of matrices before plotting is
     recommended.
}

