% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.prior.zhang.R
\name{make.prior.zhang}
\alias{make.prior.zhang}
\title{Prior for the MCMC sampler.}
\usage{
make.prior.zhang(den.mu = "unif", par.mu = c(-100, 100), den.sd = "unif",
  par.sd = c(0, 100), unif.corr = TRUE, Sigma = NULL, nu = NULL)
}
\arguments{
\item{den.mu}{Set the density type for the phylogenetic mean. One of "unif" or "norm".}

\item{par.mu}{Numeric vector with length 2. The parameters for the density of phylogenetic means.}

\item{den.sd}{Set the density type for the vector of standard deviations. One of "unif" or "lnorm".}

\item{par.sd}{Numeric vector with length 2. The parameters for the density of standard deviations. When set to "unif", then 'par.sd[1]' (the min) need to be a positive value.}

\item{unif.corr}{Whether the}

\item{Sigma}{}

\item{nu}{}
}
\value{
List of density functions to be used as the prior for MCMC functions.
}
\description{
Generates prior densities for the MCMC sampler. Independent priors are defined for the phylogenetic mean, the vector of standard deviations and the structure of correlation. Thus allowing for a wide range of configurations. Priors for the phylogenetic mean and the standard deviations can be uniform of normal (lognormal in the case of the standard deviations). Prior on the matrix of correlations is distributed as a inverse Wishart and can be set to a marginally uniform prior or to be centered is some custom matrix.
}
\details{
For a model with two or more rate matrices fitted to the phylogeny different priors can be set for each R matrix. To do this first create a list of 'p' priors created with this function and pass the list to the MCMC sampler. The MCMC will set the prior for the phylogenetic mean (den.mu) as the prior in the first element of the list. The vector of standard deviations and the prior on the correlation will be defined as the other elements of the list in the same order as the R matrices fitted to the data.
}

