% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMLEmap.R
\name{makeMLEmap}
\alias{makeMLEmap}
\title{Map ML ancestral estimates to the tree.}
\usage{
makeMLEmap(tree, state)
}
\arguments{
\item{tree}{A phylogenetic tree of the 'phylo' format.}

\item{state}{A vector with states to map the regimes to the tree. Names need to be species and match tree$tip.label .}
}
\value{
A 'Simmap' format tree. This can be used in subsequent functions such as multi.R.iwish.mcmc.
}
\description{
Function perform marginal maximum likelihood estimate for the state at the nodes and map the branches to the respective regimes. Transitions between regimes are made in the mid point of the branches.
}
\details{
If the tree shows politomies the function will solve the politomies randomly and add a branch of length equal to l/100, where l is the smallest branch length observed in the tree. This procedure will guarantee that there will be no zero branch length. The arbitrary branch added to the solved politomies is much smaller than the ones observed in the empirical tree, as a result it is very unlikely that the ancestral state reconstruction will be biased.

This analysis does not substitute the simulations of transitions between states provided by a stochastic map. We suggest the user to contrast results with a distribution of stochastic maps and, if possible, make analyses based on this map and samples from the stochastic map to incorporate uncertainty in ancestral state estimates.
}

