% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePriorBarnard.R
\name{makePriorBarnard}
\alias{makePriorBarnard}
\title{Create prior distributions.}
\usage{
makePriorBarnard(mu, sd, min, max)
}
\arguments{
\item{mu}{numeric. Mean value for the multivariate normal prior on the phylogenetic mean or root value.}

\item{sd}{numeric. Standard deviation for the multivariate normal prior on the phylogenetic mean.}

\item{min}{numeric. Minimum bound of the uniform prior on the variance component of the covariance matrix.}

\item{max}{numeric. Maximum bound of the uniform prior on the variance component of the covariance matrix.}
}
\value{
List of prior functions..
}
\description{
Gerate a list of prior functions for the log density of the evolutionary rate matrix and phylogenetic mean.
}
\details{
The prior for the evolutionary rate matrix uses the separation strategy (Barnard et al, 2000).
    This strategy creates a marginally uniform prior by modelling the covariance matrix using a
    correlation matrix and a vector of variance coeficients. One can also use a inverse Wishart prior,
    however the inverse Wishart is a very informative prior and need to be explored with caution.\cr
\cr
The prior on the phylogenetic mean (root value) is a multivariate normal distribution with mean vector
    'mu' and variance vector 'sd'.
}
\references{
Barnard, J., R. McCulloch, and X.-L. Meng. 2000. Modeling covariance matrices in terms of standard deviations and correlations, with application to shrinkage. Statistica Sinica 10:1281–1312.
}

