% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePriorIWish.R
\name{makePriorIWish}
\alias{makePriorIWish}
\title{Create prior distributions.}
\usage{
makePriorIWish(mu, sd, v, p, S)
}
\arguments{
\item{mu}{numeric. Mean value for the multivariate normal prior on the phylogenetic mean or root value.}

\item{sd}{numeric. Standard deviation for the multivariate normal prior on the phylogenetic mean.}

\item{v}{numeric. Degrees of freedom for the invere Wishart distribution.}

\item{p}{numeric. The dimension of the evolutionary rate matrix.}

\item{S}{matrix. The scale matrix for the inverse Wishart distribution.}
}
\value{
List of prior functions.
}
\description{
Gerate a list of prior functions for the log density of the evolutionary rate matrix and phylogenetic mean.
}
\details{
The prior for the evolutionary rate matrix uses the inverse wishart. The inverse Wishart is a very
    informative prior and need to be explored with caution. Please see 'make.prior.barnard' for a
    usually better prior option.
\cr
The prior on the phylogenetic mean (root value) is a multivariate normal distribution with mean vector
    'mu' and variance vector 'sd'.
}

