% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePropIWish.R
\name{makePropIWish}
\alias{makePropIWish}
\title{Proposal for covariance matrices.}
\usage{
makePropIWish(curr.vcv, k, v)
}
\arguments{
\item{curr.vcv}{matrix. Current state of the covariance matrix.}

\item{k}{numeric. Number of dimensions of the matrix.}

\item{v}{numeric. Degrees of freedom for the inverse wishart distribution.}
}
\value{
A matrix.
}
\description{
make a proposal for a new covariance matrix given the current state.
}
\details{
Internal function to be used in the MCMC.
}

