% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePropMean.R
\name{makePropMean}
\alias{makePropMean}
\title{Phylogenetic mean proposal and accept/reject.}
\usage{
makePropMean(cache.data, cache.chain, prior, w_sd, w_mu, v, iter, count,
  traitwise)
}
\arguments{
\item{cache.data}{The cache for the data.}

\item{cache.chain}{The cache with the MCMC chain.}

\item{prior}{List with prior functions.}

\item{w_sd}{}

\item{w_mu}{}

\item{v}{The degrees of freedom parameter for the inverted-wishart distribution.}

\item{iter}{Tracks the current generation of the MCMC chain.}

\item{count}{Used to track the accept and reject steps of the MCMC.}

\item{traitwise}{}

\item{w}{The width parameter for the sliding-window proposal step of the phylogenetic mean.}
}
\value{
Updated version of the cache.chain.
}
\description{
Make the proposal and accept/reject for the phylogenetic mean.
}
\details{
Internal function to be used with the MCMC.
}

