% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePropMultSigma.R
\name{makePropMultSigma}
\alias{makePropMultSigma}
\title{Sigma step using Zhang strategy}
\usage{
makePropMultSigma(cache.data, cache.chain, prior, v, w_sd, w_mu, iter, count,
  files)
}
\arguments{
\item{cache.data}{}

\item{cache.chain}{}

\item{prior}{}

\item{iter}{}

\item{count}{}

\item{w}{}
}
\value{
The chain cache.
}
\description{
Description!
}
\details{
Details!
}
\author{
daniel
}

