% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePropMultSigmaList.R
\name{makePropMultSigmaList}
\alias{makePropMultSigmaList}
\title{Sigma step using Zhang strategy}
\usage{
makePropMultSigmaList(cache.data, cache.chain, prior, v, w_sd, w_mu, iter,
  count, files, n.phy)
}
\arguments{
\item{cache.data}{}

\item{cache.chain}{}

\item{prior}{}

\item{v}{}

\item{w_sd}{}

\item{w_mu}{}

\item{iter}{}

\item{count}{}

\item{n.phy}{}
}
\value{
The chain cache.
}
\description{
Description!
}
\details{
Details!
}
\author{
daniel
}

