% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePropSingleSigmaList.R
\name{makePropSingleSigmaList}
\alias{makePropSingleSigmaList}
\title{Sigma step using Zhang strategy}
\usage{
makePropSingleSigmaList(cache.data, cache.chain, prior, w_sd, w_mu, v, iter,
  count, files, phy, n.phy)
}
\arguments{
\item{cache.data}{Description}

\item{cache.chain}{Description}

\item{prior}{Description}

\item{w_sd}{Description}

\item{w_mu}{Description}

\item{v}{Description}

\item{iter}{Description}

\item{count}{Description}
}
\value{
The chain cache.
}
\description{
Description!
}
\details{
Details!
}

