% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStartSeparation.R
\name{makeStartSeparation}
\alias{makeStartSeparation}
\title{Start point for MCMC.}
\usage{
makeStartSeparation(n.traits, n.R)
}
\arguments{
\item{n.traits}{Number of traits in the model.}

\item{n.R}{Number of R matrices fitted to the tree.}
}
\value{
A list with the starting points for a MCMC. This function will draw from an uniform prior. Same as created by 'make.prior.zhang'.
}
\description{
Create a list with the starting points for the MCMC.
}
\details{
Create a list with the starting points for the MCMC.
}

