% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePosterior.R
\name{mergePosterior}
\alias{mergePosterior}
\title{Merge posterior distributions}
\usage{
mergePosterior(...)
}
\arguments{
\item{...}{Any number of posterior distributions as produced by the function 'readMCMC'.}
}
\value{
A merged posterior distribution in the same format.
}
\description{
Join two or more independent MCMC chains from the same data and phylogenetic trees by appending them together into a single chain.
}
\examples{
## Run two (very short) chains and merge results.
data(centrarchidae)
handle1 <- ratematrixMCMC(data=centrarchidae$data, phy=centrarchidae$phy.map, gen=1000)
posterior1 <- readMCMC(handle1, burn=0.25, thin=1)
handle2 <- ratematrixMCMC(data=centrarchidae$data, phy=centrarchidae$phy.map, gen=1000)
posterior2 <- readMCMC(handle2, burn=0.25, thin=1)
merged <- mergePosterior(posterior1, posterior2)
merged ## Summary of the chain
plotRatematrix(merged)
}
\author{
Daniel S. Caetano and Luke J. Harmon
}

