% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSimmap.R
\name{mergeSimmap}
\alias{mergeSimmap}
\title{Merge two or more regimes of a 'simmap' tree.}
\usage{
mergeSimmap(phy, merge.regimes = NULL, new.regime = NULL,
  drop.regimes = FALSE)
}
\arguments{
\item{phy}{a phylogeny of the 'simmap' format.}

\item{merge.regimes}{a vector with the names of the regimes to be merged.}

\item{new.regime}{the name of the new regime}

\item{drop.regimes}{whether to simply drop all information about the regimes and return a phylogeny of class 'phylo'.}
}
\value{
A phylogeny of the format 'simmap' with merged regimes.
}
\description{
Function will merge regimes to form a new regime. The distribution of the regimes across the tree will not change. The function only re-label the regimes such that two or more regime are now recognized as a unique new regime. The function can also drop all information about regimes and return a phylogeny of the class 'phylo'.
}
\details{
Function takes the elements of 'merge.regimes' and collapse all those regimes into a single one. The branch length associated with 'merge.regimes' are summed and assigned to the regime correspondent to the first element of the 'merge.regimes' vector. Then this new regime is renamed as 'new.regime'.\cr
\cr
The original distribution of the regimes across the branches of the phylogeny will not change. This function only manipulate the labels of the regimes.\cr
\cr
If the original phylogeny has only two regimes or if 'drop.regimes' is set to TRUE, then function will drop all the information about regimes and the phylogeny will be of class 'phylo' only.
}
\author{
daniel
}

