% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.phylo.mean.step.fast.R
\name{multi.phylo.mean.step.fast}
\alias{multi.phylo.mean.step.fast}
\title{Proposal and accept/reject for the phylogenetic mean.}
\usage{
multi.phylo.mean.step.fast(cache.data, cache.chain, prior, v, w_sd, w_mu, iter,
  count)
}
\arguments{
\item{cache.data}{list. The cache with the data.}

\item{cache.chain}{list. The cache with the MCMC chain.}

\item{prior}{list. A list with the prior functions.}

\item{v}{numeric. Degrees of freedom for the inverse-wishart distribution.}

\item{iter}{numeric. The state of the MCMC chain. This is used to access elements in the MCMC caches.}

\item{count}{numeric. Keep the count of the chain to record accepted and rejected steps.}

\item{w}{numeric. The sliding window width parameter.}
}
\value{
Return a modified 'cache.chain'.
}
\description{
Make the proposal and acceptance steps for the phylogenetic mean.
}
\details{
Internal function to be used in the MCMC. This functions uses a simple sliding window strategy to perform the updates.
}

