% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R.iwish.mcmc.R
\name{multi.R.iwish.mcmc}
\alias{multi.R.iwish.mcmc}
\title{MCMC for two or more evolutionary rate matrices.}
\usage{
multi.R.iwish.mcmc(X, phy, start, prior, gen, v, w, prop = c(0.3, 0.7), chunk,
  dir = NULL, outname = "single_R_fast", IDlen = 5)
}
\arguments{
\item{X}{matrix. A matrix with the data. 'rownames(X) == phy$tip.label'.}

\item{phy}{simmap phylo. A phylogeny of the class "simmap" from the package 'phytools'. Function uses the location information for a number of traits equal to the number of fitted matrices.}

\item{start}{list. Element [[1]] is the starting value for the phylogenetic mean and element [[2]] is the starting value for the R matrices. Element [[2]] is also a list with length equal to the number of matrices to be fitted to the data.}

\item{prior}{list. Produced by the output of the function 'make.prior.barnard' or 'make.prior.diwish'. First element of the list [[1]] is a prior function for the log density of the phylogenetic mean and the second element [[2]] is a prior function for the evolutionary rate matrix (R). The prior can be shared among the rate matrices or be set a different prior for each matrix. At the moment the function only produces a shared prior among the fitted matrices. Future versions will implement independent priors for each of the fitted matrices.}

\item{gen}{numeric. Number of generations of the MCMC.}

\item{v}{numeric. Degrees of freedom parameter for the inverse-Wishart proposal distribution for the evolutionary rate matrix.}

\item{w}{numeric. Width of the sliding window proposal step for the phylogenetic mean.}

\item{prop}{vector. The proposal frequencies. Vector with two elements (each between 0 and 1). First is the probability that the phylogenetic mean will be sampled for a proposal step at each genetarion, second is the probability that the evolutionary rate matrix will be updated instead. First the function sample whether the root value or the matrix should be updated. If the matrix is selected for an update, then one of the matrices fitted to the phylogeny is selected to be updated at random with the same probability.}

\item{chunk}{numeric. Number of generations that the MCMC chain will be stored in memory before writing to file. At each 'chunk' generations the function will write the block stored in memory to a file and erase all but the last generation, which is used to continue the MCMC chain. Each of the covariance matrices is saved to its own file.}

\item{dir}{string. Directory to write the files, absolute or relative path. If 'NULL' then output is written to the directory where R is running (see 'getwd()'). If a directory path is given, then function will test if the directory exists and use it. If directiory does not exists the function will try to create one.}

\item{outname}{string. Name pasted to the files. Name of the output files will start with 'outname'.}

\item{IDlen}{numeric. Set the length of the unique numeric identifier pasted to the names of all output files. This is set to prevent that multiple runs with the same 'outname' running in the same directory will be lost.Default value of 5 numbers, something between 5 and 10 numbers should be good enough. IDs are generated randomly using the function 'sample'.}
}
\value{
Fuction creates files with the MCMC chain. Each run of the MCMC will be identified by a unique identifier to facilitate identification and prevent the function to overwrite results when running more than one MCMC chain in the same directory. See argument 'IDlen'. The files in the directory are: 'outname.ID.loglik': the log likelihood for each generation, 'outname.ID.n.matrix': the evolutionary rate matrix n, one per line. Function will create one file for each R matrix fitted to the tree, 'outname.ID.root': the root value, one per line. \cr
\cr
Additionally it returns a list object with information from the analysis to be used by other functions. This list is refered as the 'out' parameter in those functions. The list is composed by: 'acc_ratio' numeric vector with 0 when proposal is rejected and non-zero when proposals are accepted. 1 indicates that root value was accepted, 2 and higher indicates that the first or subsequent matrices were updated; 'run_time' in seconds; 'k' the number of matrices fitted to the tree; 'p' the number of traits in the analysis; 'ID' the identifier of the run; 'dir' directory were output files were saved; 'outname' the name of the chain, appended to the names of the files; 'trait.names' A vector of names of the traits in the same order as the rows of the R matrix, can be used as the argument 'leg' for the plotting function 'make.grid.plot'; 'data' the original data for the tips; 'phy' the phylogeny; 'prior' the list of prior functions; 'start' the list of starting parameters for the MCMC run; 'gen' the number of generations of the MCMC.
}
\description{
Runs a Markov chain Monte Carlo (MCMC) chain to estimate the posterior distribution of two or more
   evolutionary rate matrices (R) fitted to the phylogeny.
}
\details{
MCMC using the inverse-Wishart as a proposal distribution for the covariance matrix and a simple sliding
   window for the phylogenetic mean in the random walk Metropolis-Hastings algorithm. At the moment the
   function only applies the 'rpf' method for calculation of the log likelihood implemented in the
   package 'mvMORPH'. Future versions should offer different log likelihood methods for the user.
This version is using the pruning algoritm to c
}

