% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiR.matrix.step.zhang.R
\name{multi.sigma.step.zhang}
\alias{multi.sigma.step.zhang}
\title{Sigma step using Zhang strategy}
\usage{
multi.sigma.step.zhang(cache.data, cache.chain, prior, v, w_sd, w_mu, iter,
  count)
}
\arguments{
\item{cache.data}{}

\item{cache.chain}{}

\item{prior}{}

\item{iter}{}

\item{count}{}

\item{w}{}
}
\value{
The chain cache.
}
\description{
Description!
}
\details{
Details!
}
\author{
daniel
}

