% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiR.loglik.R
\name{multiR.loglik}
\alias{multiR.loglik}
\title{Log likelihood for multiple matrices.}
\usage{
multiR.loglik(X, root, R.m, C.m, D, n, r, p)
}
\arguments{
\item{X}{matrix. The matrix of trait data.}

\item{root}{numeric. The phylogenetic mean.}

\item{R.m}{list. List of matrices fitted to the tree with length equal to 'cache.data$p'.}

\item{C.m}{list. List of C matrices with length equal to 'cache.data$p'. This is calculated by the function 'multiC' of the 'phytools' package.}

\item{D}{matrix. The design matrix. This is a fixed term, dependent on the phylogeny.}

\item{n}{numeric. Number of tips in the phylogeny.}

\item{r}{numeric. Number of traits, the dimension of the R matrix.}

\item{p}{numeric. Number of R matrices fitted to the tree.}

\item{method}{string. Method to be used to calculate the likelihood. Right now only the option 'rpf' is working. More options will be implemented in the future.}
}
\value{
numeric. The log likelhood.
}
\description{
Log likelihood function for two or more evolutionary rate matrices.
}
\details{
Internal function to be used by the MCMC. This function uses the 'rpf' method to calculate the approximation of the inverse of matrices as impplemented in mvMORPH.
}
\references{
Revell, L. J., and D. C. Collar. 2009. Phylogenetic Analysis of the Evolutionary Correlation Using Likelihood. Evolution 63:1090–1100.
}

