% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiR.matrix.step.fast.R
\name{multiR.matrix.step.fast}
\alias{multiR.matrix.step.fast}
\title{Proposal and accept/reject for two or more evolutionary rate matrices.}
\usage{
multiR.matrix.step.fast(cache.data, cache.chain, prior, w, v, iter, count)
}
\arguments{
\item{cache.data}{list. Cache with the data for the MCMC}

\item{cache.chain}{list. Cache with the states of the MCMC chain}

\item{prior}{list. List of prior functions.}

\item{w}{numeric. Width of the sliding-window proposal for the phylogenetic mean.}

\item{v}{numeric. Degrees of freedom for the inverse-Wishart distribution.}

\item{iter}{numeric. Gives the present state (generation) of the MCMC chain.}

\item{count}{numeric. Used to track the accept and reject steps of the MCMC chain.}
}
\value{
Updated state of the cache.chain list.
}
\description{
Makes the proposal and accept/reject steps for two or more fitted evolutionary rate matrices.
}
\details{
Internal function to be used by the MCMC
}

