% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvLogLik.R
\name{mvLogLik}
\alias{mvLogLik}
\title{Multivariate Brownian motion log likelihood}
\usage{
mvLogLik(X, phy, R, mu)
}
\arguments{
\item{X}{A matrix with the trait data.}

\item{phy}{The phylogeny with "named" regimes in the 'simmap' format. See 'phytools:make.simmap' function.}

\item{R}{A named list of rate matrices. Names must match the regimes of the 'simmap' tree.}

\item{mu}{A vector of phylogenetic means.}
}
\value{
Returns the log likelihood.
}
\description{
Calculates the log likelihood of a tree, rates and phylogenetic mean. Applies to both a single and multiple regimes. Right now it only works with a constant regime or with a regime with two rate matrices.
}
\details{
This function uses the pruning algorithm. This avoids the calculation of the inverse and the determinant of large matrices. Also it makes calculation much more stable, so the likelihood can be calculated even for very large matrices. This function is to be used outside of the MCMC, since one can precaulculate some quantities that are constant given the tree and the data.
The regimes in the 'simmap' tree need to be named to assume the proper match between the rate matrices and the tree.
}

