% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseResults.R
\name{parseResults}
\alias{parseResults}
\title{Parse simulations results.}
\usage{
parseResults(out.one.matrix, out.two.matrix, rds = FALSE, flag = NULL,
  burn = 0.5, thin = 100)
}
\arguments{
\item{out.one.matrix}{Output from MCMC}

\item{out.two.matrix}{Output from MCMC}

\item{rds}{Posterior distribution.}

\item{flag}{A name for the run.}

\item{burn}{The burnin to be excluded.}

\item{thin}{The thinning.}
}
\value{
Plots and etc.
}
\description{
Function to read and parse the results from the simulations.
}
\details{
Function to read and parse the results from the simulations. At this point the function will only work
  when there is two chains. This need to be extended and fixed for it to run with one or two chains.
}

