% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passed.heidel.R
\name{passed.heidel}
\alias{passed.heidel}
\title{Heigel diagnostic of convergence.}
\usage{
passed.heidel(mcmc)
}
\arguments{
\item{mcmc}{An 'mcmc' class object as defined in the 'coda' package.}
}
\value{
A vector of length 2 with TRUE or FALSE dependent if the test passed.
}
\description{
Check the convergence of the variables using a single chain.
}
\details{
Function applies the 'heidel.diag' test from the package 'coda'. For a better test of convergence need to use more than one chain and apply Gelman's R diagnostic.
}

