% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo.mean.step.fast.R
\name{phylo.mean.step.fast}
\alias{phylo.mean.step.fast}
\title{Phylogenetic mean proposal and accept/reject.}
\usage{
phylo.mean.step.fast(cache.data, cache.chain, prior, w, v, iter, count)
}
\arguments{
\item{cache.data}{The cache for the data.}

\item{cache.chain}{The cache with the MCMC chain.}

\item{prior}{List with prior functions.}

\item{w}{The width parameter for the sliding-window proposal step of the phylogenetic mean.}

\item{v}{The degrees of freedom parameter for the inverted-wishart distribution.}

\item{iter}{Tracks the current generation of the MCMC chain.}

\item{count}{Used to track the accept and reject steps of the MCMC.}
}
\value{
Updated version of the cache.chain.
}
\description{
make the proposal and accept/reject for the phylogenetic mean.
}
\details{
Internal function to be used with the MCMC.
}

