% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRootValue.R
\name{plotRootValue}
\alias{plotRootValue}
\title{Plot the posterior distribution of the root values for the traits.}
\usage{
plotRootValue(chain, color = "black", set.xlab = NULL, set.cex.lab = 1,
  set.cex.axis = 1.5, set.xlim = NULL, hpd = 100, mfrow = 1,
  vline.values = NULL, vline.color = NULL, vline.wd = NULL,
  show.zero = FALSE)
}
\arguments{
\item{chain}{the posterior distribution loaded from the files using 'readMCMC'.}

\item{color}{the color for the histograms.}

\item{set.xlab}{a vector with legends for the x axes. If 'NULL' (default), the names are 'trait_1' to 'trait_n".}

\item{set.cex.lab}{the cex value for the labels (default is 1).}

\item{set.cex.axis}{the cex value for the axes numbers (default is 1.5).}

\item{set.xlim}{the xlim for the plot. Need to be a vector with the lower and higher bound.}

\item{hpd}{the Highest Posterior Density interval to highlight in the plot. Regions outside the interval will be painted in "white". A numeric value between 0 and 100 (default is 100).}

\item{mfrow}{the number of lines for the plate. The plots will be arranged in the number of lines provided (default is 1).}

\item{vline.values}{numeric values for a vertical line in the plot. Can be a single value recycled for each of the plots or a vector with length equal to the number of traits.}

\item{vline.color}{vector with colors for the vertical lines. Can be a single color if length of 'vline.values' is 1 otherwise need to have length equal to the number of traits.}

\item{vline.wd}{numeric value for the width of the vertical lines. Can be a single value if length of 'vline.values' is 1 otherwise need to have length equal to the number of traits.}

\item{show.zero}{whether a vertical line should be plotted showing the position of the value 0 in the plot.}
}
\value{
A plot with the posterior density for the root values of the MCMC analysis. Many characteristics of the plot can be controlled by the user.
}
\description{
Plot the posterior distribution of root values sampled from the MCMC analysis using 'ratematrixMCMC' function.
}
\examples{
\donttest{
## Load data
data(centrarchidae)
## Run MCMC. This is just a very short chain.
handle <- ratematrixMCMC(data=centrarchidae$data, phy=centrarchidae$phy.map, gen=1000)
## Load posterior distribution, make plots and check the log.
posterior <- readMCMC(handle, burn=0.25, thin=1)
plotRatematrix(posterior)
plotRootValue(posterior)
plotPrior(handle)
plotPrior(handle, root=TRUE)
logAnalizer(handle)
}
}
\author{
Daniel S. Caetano and Luke J. Harmon
}

