% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postPredicSims.R
\name{postPredicSims}
\alias{postPredicSims}
\title{Make posterior predictive simulations.}
\usage{
postPredicSims(mcmc.chain, out, n.samples = 100, mc.cores = NULL)
}
\arguments{
\item{mcmc.chain}{The posterior chain generated by 'read.single.R.iwish', 'read.multi.R.iwish', or 'mergePosterior'.}

\item{out}{The 'out' file from the MCMC function.}

\item{n.samples}{Number of samples to be taken from the posterior distribution.}

\item{mc.cores}{The number of cores to be used for parallel simulations. Default value is 'NULL' and uses a single computer processor.}
}
\value{
A list of matrices. Each element is a simulation of datasets. Using the same model and parameter estimates from the joint posterior distribution.
}
\description{
Perform posterior predictive simulations.
}
\details{
This function performs posterior predictive simulations by sampling 'n.samples' times from the joint posterior distribution and creating a new dataset for each sample. The simulations can be used to check if the model and parameter estimates can generate data that is close to your data. This is a measure of model adequacy.
}
\seealso{
\code{read.single.R.iwish} or \code{read.multi.R.iwish} or \code{mergePosterior}
}

