% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ratematrix_multi_chain.R
\name{print.ratematrix_multi_chain}
\alias{print.ratematrix_multi_chain}
\title{Print method for the "ratematrix_multi_chain" class.}
\usage{
\method{print}{ratematrix_multi_chain}(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Additional arguments. Not used here.}
}
\description{
Print method for the "ratematrix_multi_chain" class.
}
\details{
Print information about object.
}
