% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ratematrix_prior_function.R
\name{print.ratematrix_prior_function}
\alias{print.ratematrix_prior_function}
\title{Print method for the "ratematrix_prior_sample" class.}
\usage{
\method{print}{ratematrix_prior_function}(x)
}
\arguments{
\item{x}{The object.}
}
\description{
Print method for the "ratematrix_prior_sample" class.
}
\details{
Print information about object.
}

