% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.vcv.barnard.R
\name{prior.vcv.barnard}
\alias{prior.vcv.barnard}
\title{Prior density under the Barnard separation.}
\usage{
prior.vcv.barnard(S, min = 0, max = 100)
}
\arguments{
\item{S}{The variance covariance matrix. The evolutionary rate matrix.}

\item{min}{Minimum of the uniform prior on the variance component.}

\item{max}{Maximum of the uniform prior on the variance component.}
}
\value{
The log density of the matrix S under the prior.
}
\description{
Function to calculate the prior density using the Barnard separation.
}
\details{
Internal function to be used by the MCMC.
}

