% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.matrix.step.fast.R
\name{R.matrix.step.fast}
\alias{R.matrix.step.fast}
\title{Proposal and accept reject of the evolutionary rate matrix.}
\usage{
R.matrix.step.fast(cache.data, cache.chain, prior, w, v, iter, count)
}
\arguments{
\item{cache.data}{Cache of data.}

\item{cache.chain}{Cache of the MCMC chain}

\item{prior}{List with the prior functions}

\item{w}{Width of the sliding-window parameter}

\item{v}{Degrees of freedom parameter for the inverted-wishart.}

\item{iter}{Records the generation of the MCMC chain.}

\item{count}{Keep track of the accept reject.}
}
\value{
Updated cache.chain.
}
\description{
Makes the proposal and acceptance step for the R matrix.
}
\details{
Internal function to be used with the MCMC.
}

