% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.chart.R
\name{rate.chart}
\alias{rate.chart}
\title{Histograms of the rates.}
\usage{
rate.chart(mat, burn = 0.5, thin = 1, true.R = NULL, file = NULL)
}
\arguments{
\item{mat}{List of covariance matrices.}

\item{burn}{proportion of burnin}

\item{thin}{The thinning.}

\item{true.R}{A point estimate to be included in the plots. This will always be a vertical line in 'red'.}

\item{file}{When a file string is provided the function creates a jpeg figure file istead of creating the plot in the standard device.}
}
\value{
Plot in the standard device or a 'jpeg' file.
}
\description{
Make histograms of the rates for the evolutionary rate matrix.
}
\details{
Internal function to be used with 'make.grid.plot'.
}

