% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratematrix-package.R
\docType{package}
\name{ratematrix}
\alias{ratematrix}
\alias{ratematrix-package}
\title{ratematrix.}
\description{
R package to estimate the evolutionary rate matrix using Bayesian Markov-chain Monte Carlo for multiple regimes.
}
