% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratematrixMCMC.R
\name{ratematrixMCMC}
\alias{ratematrixMCMC}
\title{Run the MCMC chain for the evolutionary rate matrix model.}
\usage{
ratematrixMCMC(data, phy, prior = "empirical_mean", start = "prior_sample",
  gen, v = 50, w_sd = 0.5, w_mu = 0.5, prop = c(0.025, 0.975),
  chunk = gen/100, dir = NULL, outname = "ratematrixMCMC", IDlen = 5,
  singlerate = FALSE)
}
\arguments{
\item{data}{a matrix with the data. Each column is a different trait and species names need to be provided as rownames (rownames(data) == phy$tip.label).}

\item{phy}{a phylogeny of the class "simmap" with the mapped regimes. The number of evolutionary rate matrices fitted to the phylogeny is equal to the number of regimes in phy. Regime names will also be used. See 'Details'.}

\item{prior}{the prior densities for the MCMC. Must be one of (i) "uniform", (ii) "empirical_mean" (the default), (iii) the output of the "makePriorSeparation" function or (iv) a list of functions. See 'Details'.}

\item{start}{the starting state for the MCMC chain. Must be one of (i) "prior_sample" (the default), (ii) "mle", (iii) a list object. See 'Details'.}

\item{gen}{number of generations for the chain.}

\item{v}{value for the degrees of freedom parameter of the inverse-Wishart proposal distribution for the correlation matrix.}

\item{w_sd}{value for the width of the uniform proposal distribution for the vector of standard deviations.}

\item{w_mu}{value for the width of the uniform proposal distribution for the vector of root values (phylogenetic mean).}

\item{prop}{the proposal frequencies for each parameter of the model (default is 'c(0.025,0.975)'). This needs to be a numeric vector of length 2. Each value need to be between 0 and 1 and the sum of the vector equal to be 1. These values are the probability that the phylogenetic mean or the set of evolutionary rate matrices will be updated at each step of the MCMC chain, respectively.}

\item{chunk}{number of generations that the MCMC chain will be kept in the computer memory before writing to file. Larger values will use more RAM memory.}

\item{dir}{path of the directory to write the files (default is 'NULL'). If 'NULL' then function will write files to current directory (check 'getwd()'). If directory does not exist, then function will create a new directory. If function does not have permission (or fail) to write files to the pointed directory, then program will stop with an error message.}

\item{outname}{name for the MCMC chain (default is 'ratematrixMCMC'). Name will be used in all the files alongside a unique ID of numbers with length of 'IDlen'.}

\item{IDlen}{length of digits of the numeric identifier used to name output files (default is 5).}

\item{singlerate}{whether the function should fit a single regime and ignore the number of regimes painted to the tree (default is FALSE).}
}
\value{
Function returns a list with the details of the MCMC run and write the MCMC output to the directory in a series of files.
}
\description{
Function runs a MCMC chain to estimate the posterior distribution of the evolutionary rate matrix and the phylogenetic root value given a phylogeny and trait data. Prior distribution and starting state of the chain can be chosen from one of the options provided. The function also accepts custom priors and starting states. Note that function requires output files in the chosen directory to be created and accessed. See 'Details' and 'Examples' for more information.
}
\details{
Talk about function 'x' to collapse two or more regimes of a 'simmap' class tree.
Explain the structure of the prior so that people can build their own.
Explain how to fix the possible problem of function cannot write to the current directory. Need to be something simple. Maybe bullet proff it somehow.\cr
\cr
Fuction creates files with the MCMC chain. Each run of the MCMC will be identified by a unique identifier to facilitate identification and prevent the function to overwrite results when running more than one MCMC chain in the same directory. See argument 'IDlen'. The files in the directory are: 'outname.ID.loglik': the log likelihood for each generation, 'outname.ID.n.matrix': the evolutionary rate matrix n, one per line. Function will create one file for each R matrix fitted to the tree, 'outname.ID.root': the root value, one per line. \cr
\cr
Additionally it returns a list object with information from the analysis to be used by other functions. This list is refered as the 'out' parameter in those functions. The list is composed by: 'acc_ratio' numeric vector with 0 when proposal is rejected and non-zero when proposals are accepted. 1 indicates that root value was accepted, 2 and higher indicates that the first or subsequent matrices were updated; 'run_time' in seconds; 'k' the number of matrices fitted to the tree; 'p' the number of traits in the analysis; 'ID' the identifier of the run; 'dir' directory were output files were saved; 'outname' the name of the chain, appended to the names of the files; 'trait.names' A vector of names of the traits in the same order as the rows of the R matrix, can be used as the argument 'leg' for the plotting function 'make.grid.plot'; 'data' the original data for the tips; 'phy' the phylogeny; 'prior' the list of prior functions; 'start' the list of starting parameters for the MCMC run; 'gen' the number of generations of the MCMC.
}
\author{
daniel
}

