% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSingleRegimeMCMC.R
\name{readSingleRegimeMCMC}
\alias{readSingleRegimeMCMC}
\title{Read output files of MCMC with a single R matrix.}
\usage{
readSingleRegimeMCMC(out, burn = 0.5, thin = 1, dir = NULL)
}
\arguments{
\item{out}{list. The list object returned by the function 'single.R.iwish.mcmc'. See more details in 'help(single.R.iwish.mcmc)'.}

\item{burn}{numeric. The proportion of the burnin to be pruned from the MCMC chain.}

\item{thin}{numeric. The thinning of the posterior distribution. Same format as the argument 'by' of the function 'seq'.}

\item{dir}{string. Directory where output files from the MCMC run are stored. If 'NULL', then function tries to read the files from the current directory. This can be set to a directory different from the 'dir' argument of the 'single.R.iwish.mcmc' function.}
}
\value{
List with the MCMC chain for the phylogenetic mean (root value) and evolutionary rate matrices (R). 'root' are the values for the phylogenetic mean; 'matrix' is a list with the chain of R matrices; 'log.lik' is the log likelihood (not posterior) for the chain.
}
\description{
Read the results of the MCMC for a single evolutionary rate (R) matrix.
}
\details{
Function will use 'readr' package to read the output files produced by the Markov chain Monte Carlo
   analysis made with the function 'single.R.iwish.mcmc'.
}

