% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rinvchisq.R
\name{rinvchisq}
\alias{rinvchisq}
\title{Random sample from scaled inverse chi-squared distribution.}
\usage{
rinvchisq(n = 1, v.chi = 1, scale = 1/v.chi)
}
\arguments{
\item{n}{numeric. Number of samples.}

\item{v.chi}{numeric. Degrees of freedom.}

\item{scale}{numeric. The scale factor.}
}
\value{
numeric. Sample from the distribution.
}
\description{
Generate samples from an scaled inverse chi-squared distribution.
}
\details{
Internal function to be used by the MCMC.
}

