% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riwish.R
\name{riwish}
\alias{riwish}
\title{Sample from inverse-Wishart.}
\usage{
riwish(v, S)
}
\arguments{
\item{v}{numeric. Degrees of freedom.}

\item{S}{matrix. The scale matrix.}
}
\value{
matrix. Returns a sample from a inverse-Wishart distribution.
}
\description{
Generates a random draw from a inverse-Wishart distribution.
}

