% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundoff.R
\name{roundoff}
\alias{roundoff}
\title{Round data.}
\usage{
roundoff(a, digits, up)
}
\arguments{
\item{a}{numeric. Quantity to be rounded.}

\item{digits}{numeric. Number of digits.}

\item{up}{numeric. Control parameter.}
}
\value{
The rounded quantity.
}
\description{
A custom function to round data.
}
\details{
Custom function to round data. Internal function to be used by 'vcv.ellipse'.
}

