% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.prior.zhang.R
\name{sample.prior.zhang}
\alias{sample.prior.zhang}
\title{Sample from the prior}
\usage{
sample.prior.zhang(n, prior, sample.sd = TRUE)
}
\arguments{
\item{n}{Numeric. Number of samples from the prior.}

\item{prior}{List. The output from the "make.prior.zhang" function.}

\item{sample.sd}{Whether the function should sample the vector of standard deviations independently from the samples of the covariance matrices. If set to FALSE then the standard deviations will be derived from the covariance matrices. If set to TRUE (default) then standard deviations are independent from the covariance matrices and are sampled from their correspondent prior density. See 'Details'.}
}
\value{
A list with the parameters draw from the prior. The structure is the same to be used as the "start" parameter of the MCMC chain.
}
\description{
Make samples from the prior density.
}
\details{
The function will inherit the same parameters from "make.prior.zhang" function.

This function is useful to generate a starting value for a MCMC chain and to plot the prior distribution for the model. When generating the starting value make sure to set n=1. When plotting the prior distribution, one need to use a larger number of samples to have a nice looking plot. For the starting value of a MCMC one can choose whether to sample the starting vector of standard deviations from its prior or to derive from the covariance matrix. If the MCMC is falling to run, changing "sample.sd" to FALSE might help the evaluation of the likelihood function. However, if you are taking samples to plot the prior, make sure to set "sample.sd" to TRUE, otherwise the plot will not reflect the set of prior densities used in the MCMC chain.
}

