% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleAncestralStates.R
\name{sampleAncestralStates}
\alias{sampleAncestralStates}
\title{Sample ancestral states using Gibbs sampling}
\usage{
sampleAncestralStates(phy, mcmc, n = 1)
}
\arguments{
\item{phy}{a phylogeny in 'phylo' format. Same used to estimate the posterior distribution for the ratematrix model.}

\item{mcmc}{a posterior distribution as estimated with the 'ratematrixPolytopeMCMC' function.}

\item{n}{number of Gibbs samples to be taken for each of the posterior samples.}
}
\value{
an array with the samples at the nodes.
}
\description{
This function uses Gibbs sampling to sample ancestral states conditioned on the posterior distribution for the other parameters of the model.
}
\details{
Note that instead of randomly sampling from the posterior distribution, ancestral states are sampled for each of the samples in the posterior distribution in order. For each of the sample 'n' Gibbs samples are taken.
}
\author{
Daniel Caetano
}
