% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePrior.R
\name{samplePrior}
\alias{samplePrior}
\title{Take samples from the prior distribution}
\usage{
samplePrior(n, prior, sample.sd = TRUE, rebuild.R = FALSE)
}
\arguments{
\item{n}{number of samples to be generated.}

\item{prior}{the object with the prior function. See 'makePrior' for more information.}

\item{sample.sd}{whether the function should sample the vector of standard deviations independently from the correlation matrices. If set to FALSE, the samples for standard deviation will be derived from the covariance matrices. If set to TRUE (default) then standard deviations are independently sampled from the own prior distribution and are not derived from the samples of the correlation matrix.}

\item{rebuild.R}{whether the prior sample should return a evolutionary rate matrix rather than a correlation matrix and a vector of standard deviations (default is FALSE).}
}
\value{
A list with samples from the prior distribution. The structure of this list is the same as required by the parameter 'start' of the 'ratematrixMCMC'.
}
\description{
Generates samples from the prior distribution used in the MCMC chain.
}
\details{
The function will inherit the same parameters from "make.prior.zhang" function.
}
\examples{
\donttest{
par.mu <- rbind( c(-10, 10), c(-10, 10), c(-10, 10) )
par.sd <- c(0, 10)
prior <- makePrior(r=3, p=1, par.mu=par.mu, par.sd=par.sd)
## Sample prior. Standard deviation sampled independently.
samples.ind <- samplePrior(n=1000, prior=prior, sample.sd=TRUE)
## Now plot.
plotRatematrix(samples.ind, set.xlim=c(-100,100))
## Sample prior. Now standard deviation is NOT sampled, just derived from the vcv matrices
##    sampled from the inverse-Wishart distribution.
samples.cor <- samplePrior(n=1000, prior=prior, sample.sd=FALSE)
## Note how the inverse-Wishart is much more informative than the 'separation strategy'.
##    The x axis is the same.
plotRatematrix(samples.cor, set.xlim=c(-100,100))

## We can also use the sample from the prior to start the MCMC:
par.mu <- rbind( c(-10, 10), c(-10, 10) )
par.sd <- rbind( c(0, 10), c(0, 10) )
prior <- makePrior(r=2, p=2, par.mu=par.mu, par.sd=par.sd)
## Need to take a single sample to be used as the starting point.
start <- samplePrior(n=1, prior=prior)
data(centrarchidae)
handle <- ratematrixMCMC(data=centrarchidae$data, phy=centrarchidae$phy.map, gen=1000, start=start)
}
}
\author{
Daniel S. Caetano and Luke J. Harmon
}
