% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePriorBarnard.R
\name{samplePriorBarnard}
\alias{samplePriorBarnard}
\title{Samples from the covariance prior.}
\usage{
samplePriorBarnard(k, min = 0, max = 100)
}
\arguments{
\item{k}{numeric. dimension of the matrix.}

\item{min}{numeric. minimum value for the uniform prior on the variance.}

\item{max}{numeric. maximum value for the uniform prior on the variance.}
}
\value{
Returns a covariance matrix.
}
\description{
Sample a variance covariance matrix from the prior.
}
\details{
This function samples a variance covariance matrix from the prior distribution using the Barnard et al (2009) strategy to separate the matrix in its variance and correlation component. This function is used to initiate a Markov chain Monte Carlo run by sampling a evolutionary rate matrix from its prior.
}

