% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePriorSeparation.R
\name{samplePriorSeparation}
\alias{samplePriorSeparation}
\title{Sample from the prior}
\usage{
samplePriorSeparation(n, prior, sample.sd = TRUE)
}
\arguments{
\item{n}{Numeric. Number of samples from the prior.}

\item{prior}{List. The output from the "make.prior.zhang" function.}

\item{sample.sd}{Whether the function should sample the vector of standard deviations independently from the samples of the covariance matrices. If set to FALSE then the standard deviations will be derived from the covariance matrices. If set to TRUE (default) then standard deviations are independent from the covariance matrices and are sampled from their correspondent prior density. See 'Details'.}
}
\value{
A list with the parameters draw from the prior. The structure is the same to be used as the "start" parameter of the MCMC chain.
}
\description{
Make samples from the prior density.
}
\details{
The function will inherit the same parameters from "make.prior.zhang" function.

This function is useful to generate a starting value for a MCMC chain and to plot the prior distribution for the model. When generating the starting value make sure to set n=1. To plot the prior distribution you should use a larger sample size to result in a nice-looking plot. For the starting value of a MCMC one can choose whether to sample the starting vector of standard deviations from its prior or to derive from the covariance matrix. Empirical trials show that the MCMC will start easier (the likelihood will behave better) if the vector of standard deviations is computed from the sampled variance covariance matrices rather than sampled independently from the respective prior densities (set sample.sd=FALSE). If using 'sample.sd=TRUE' the MCMC might have dificulties to accept moves related to the evolutionary rate matrices. However, both options are correct for starting the MCMC, it is just a practical matter.
}

