% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.and.corr.R
\name{sd.and.corr}
\alias{sd.and.corr}
\title{Matrix to correlation and standard deviation.}
\usage{
sd.and.corr(mat, burn = 0.5, thin = 1)
}
\arguments{
\item{mat}{list. A list of posterior distribution of evolutionary rate matrices (R) or covariance matrices.}

\item{burn}{numeric. The proportion of the burnin to take from the chain.}

\item{thin}{thinning. The thinning to be applied to the posterior distribution of matrices. This is the same as the argument 'by' of the function 'seq'.}
}
\value{
Returns a list with two components. 'corr' is a list of the correlation matrices and 'sd' is a matrix with the standard deviations in the rows in the same order as the correlation matrices in 'corr'.
}
\description{
Decomposes a covariance matrix in the correlation and standard deviation components.
}
\details{
The function applies 'cov2cor' to calculate the correlation matrix from the covariance matrix and Takes the 'sqrt' of the diagonal elements to extract the standard deviations.
}

