% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRatematrix.R
\name{simRatematrix}
\alias{simRatematrix}
\title{Simulates multivariate trait evolution using a Brownian motion model.}
\usage{
simRatematrix(tree, vcv, anc = NULL, internal = FALSE)
}
\arguments{
\item{tree}{phylo. A phylogenetic tree in 'ape' format.}

\item{vcv}{matrix. A evolutionary rate matrix (R).}

\item{anc}{numeric. Vector of the same length as the number of traits to be simulated (same as the dimension of the 'vcv' matrix). This is used as the values for the root in the simulations. If 'NULL' then all traits have root value of 0.}

\item{internal}{boolean. Whether to return the values simulated for the nodes in the phylogeny. Default only returns the values simulated for the tips.}
}
\value{
Returns a matrix with each trait values for the tips. Traits are distributed in the rows and tips are distributed in the columns.
}
\description{
Function to simulate correlated traits under a multivariate Brownian motion model.
}
\details{
This is a function derived from 'sim.corrs' in the package 'phytools'. This version has some edits to make the simulations more efficient for this particular use. For all other application please refer to the original implementation of 'sim.corrs' in the package 'phytools'. The function uses a covariance matrix (evolutionary rate matrix R) to indicate the rates of the traits.
}

