% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTrees.R
\name{simTrees}
\alias{simTrees}
\title{Generate phylogenies with focus clades.}
\usage{
simTrees(tips, clade, sims)
}
\arguments{
\item{tips}{numeric. The total number of tips in the phylogeny.}

\item{clade}{numeric. The proportion of the total tips in the focus clade.}

\item{sims}{numeric. Number of phylogenies to be generated.}
}
\value{
A list with two elements. 'phy' is a "multiPhylo" object; 'nodes' is a list of the nodes in the same order as 'phy' pointing to the focus clade. The focus clade has the number of tips set by the argument 'clade'.
}
\description{
Sample trees that show at least one clade with a given proportion of the total diversity.
}
\details{
Function uses rejection sampling to generate phylogenetic trees with at least one clade that show a fixed proportion of the diversity. This can be used to perform simulations.
}

