% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slideWindow.R
\name{slideWindow}
\alias{slideWindow}
\title{Sliding window proposal.}
\usage{
slideWindow(x, w)
}
\arguments{
\item{x}{numeric. The current value.}

\item{w}{numeric. The width of the proposal.}
}
\value{
A updated value based on 'x'.
}
\description{
A simple sliding window proposal.
}
\details{
A simple sliding window proposal. Internal function to be used by the MCMC.
}

