% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcv.ellipse.R
\name{vcv.ellipse}
\alias{vcv.ellipse}
\title{Ellipse plots.}
\usage{
\method{vcv}{ellipse}(mat, mu, traits = c(1, 2), burn = 0.5, thin = 1,
  sample.line = 50, n.points = 200)
}
\arguments{
\item{mat}{list. A list of covariance matrices. The posterior distribution of evolutionary rate matrices fitted to the tree.}

\item{mu}{matrix. Matrix of values to be used as centroid to the ellipses. This allows the user to supply a different centroid for each trait.}

\item{traits}{vector. This is the index of the traits to be plotted. It is a length 2 numeric vector, the first indicates the row of the matrix and the second indicates the column.}

\item{burn}{numeric. Proportion of burnin to be excluded from the chain.}

\item{thin}{numeric. Thinning of the posterior chain. This is the same as the argument 'by' of the function 'seq'.}

\item{sample.line}{numeric. The number of lines to appear in the final plot. This is equal to the number of samples taken from the posterior chain.}

\item{n.points}{numeric. Number of points used to calculate the ellipse.}
}
\value{
Plot the ellipse for a combination of two traits. This function does not plot the combination of all plots in a grid. For this see 'make.grid.plot'.
}
\description{
This creates and draws a ellipse from a covariance matrix.
}
\details{
Creates ellipse plots from a covariance matrix.
}

