% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcv.proposal.plate.R
\name{vcv.proposal.plate}
\alias{vcv.proposal.plate}
\title{Unclear}
\usage{
\method{vcv}{proposal.plate}(mat, mu, burn = 0, thin, sample.line,
  n.points = 200, start = NULL, true.vcv = NULL, true.mu = NULL)
}
\arguments{
\item{mat}{list of covariance matrices.}

\item{mu}{matrix with the centroid.}

\item{burn}{proportion of burnin}

\item{thin}{thinning.}

\item{sample.line}{number of lines to be plotted.}

\item{n.points}{number of points used to calculate the ellipses.}

\item{start}{The start value used for the MCMC chain}

\item{true.vcv}{The point estimate or true value (in case of a simulation). This is going to be plotted to indicate the glocal maximum.}

\item{true.mu}{The point estimate or the true value (in case of a simulation).}
}
\value{
Plot the spread of the proposals. The matrices should start close to the start value and then gravitate toward the point estimate (or true value).
}
\description{
Only works for a 2x2 matrix.
}
\details{
Unclear
}

