% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeToMultFile.R
\name{writeToMultFile}
\alias{writeToMultFile}
\title{Write the MCMC to files.}
\usage{
writeToMultFile(files, cache.chain, p, chunk)
}
\arguments{
\item{files}{list of the files for each of the parameters.}

\item{cache.chain}{The object with the MCMC generations.}

\item{p}{the number of matrices fitted in the model.}

\item{chunk}{The interval of generations when this function should be called to write parameter states to file and clean memory.}
}
\value{
Return a updated version of cache.chain. This chain will now only have the last state in the chain. Function writes to file the content in 'cache.chain'. This function will append the new generations to the previous chunk.
}
\description{
Write the MCMC to files.
}
\details{
This function writes the chain stored in the 'cache.chain' list of parameters in the MCMC. Then the function will keep only the last generation in the chain and exclude the rest to save space in the RAM memory.
}

