% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dic.mcmc.R
\name{dic.mcmc}
\alias{dic.mcmc}
\title{Deviance Information Criteria}
\usage{
dic.mcmc(out, post)
}
\arguments{
\item{out}{The object with the MCMC info. The return from the MCMC functions.}

\item{post}{The posterior chain read by the 'read.multi.R.iwish' or 'read.single.R.iwish'.}
}
\value{
numeric. Return the DIC value.
}
\description{
Deviance information criterion (DIC) for MCMC models.
}
\details{
Function works with the samples from the R matrix MCMC estimate functions.
\cr
\cr
Important issue: DIC is prone to missbehave and reject the model with two or more matrixes even when there is clearly a better estimate of the parameters. The posterior distribution for the R matrix is not normally distributed and this might be the reason why the DIC is not a good statistics to test the models comparing different numbers of R matrices. We recommend caution while using DIC.
}
\references{
Gelman, A., J. Hwang, and A. Vehtari. 2013. Understanding predictive information criteria for Bayesian models. Stat Comput. 24:997–1016.
}
\seealso{
\code{\link{multi.R.iwish.mcmc.R single.R.iwish.mcmc.R}}
}

