% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logAnalizer.R
\name{logAnalizer}
\alias{logAnalizer}
\title{Make analysis of the log file of the MCMC.}
\usage{
logAnalizer(out, burn = 0.25, thin = 1, show.plots = TRUE,
  print.result = TRUE, dir = NULL)
}
\arguments{
\item{out}{the output object from the 'ratematrixMCMC' function.}

\item{burn}{the proportion of burn-in. A numeric value between 0 and 1.}

\item{thin}{the number of generations to skip when reading the posterior distribution.}

\item{show.plots}{whether to show a trace plot of the log-likelihood and the acceptance ratio.}

\item{print.result}{whether to print the results of the acceptance ratio to the screen.}

\item{dir}{the directory where to find the log file. If set to NULL the function will search for the files in the same directory that the MCMC chain was made.}
}
\value{
A named vector with the acceptance ratio for the whole MCMC and each parameter. If a list of phylogenetic trees was provided to the MCMC chain, then the output is a list with the acceptance ration for the parameters and a table showing the frequency in which each of the phylogenies was present when a move step was accepted.
}
\description{
Read the log file produced by the 'ratematrixMCMC' function. Calculates acceptance ration and show the trace plot.
}
\author{
daniel
}

