% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePrior.R
\name{makePrior}
\alias{makePrior}
\title{Generate prior distributions for the multivariate Brownian motion model}
\usage{
makePrior(r, p, den.mu = "unif", par.mu, den.sd = "unif", par.sd,
  unif.corr = TRUE, Sigma = NULL, nu = NULL)
}
\arguments{
\item{r}{number of traits in the model.}

\item{p}{number of evolutionary rate matrix regimes fitted to the phylogeny.}

\item{den.mu}{a string to set the density type for the phylogenetic mean. One of uniform ("unif", default) or normal ("norm") distribution.}

\item{par.mu}{the parameters for the prior density on the vector of phylogenetic means. Matrix with 2 columns and number of rows equal to the number of traits (r). When the density ('den.mu') is set to "unif" then par.mu[,1] is the minimum values and par.mu[,2] is the maximum values for each trait. When the density is set to "norm" then par.mu[,1] is the mean values and par.mu[,2] is the standard deviation values for the set of normal densities around the vector of phylogenetic means.}

\item{den.sd}{a string to set the density type for the vector of standard deviations. One of uniform ("unif", default) or log-normal ("lnorm") distribution.}

\item{par.sd}{the parameters for the density of standard deviations. Matrix with 2 columns and number of rows equal to the number of evolutionary rate matrix regimes fitted to the phylogenetic tree (p). When "den.sd" is set to "unif", then 'par.sd[,1]' (the minimum) need to be a vector of positive values and 'par.sd[,2]' is the vector of maximum values. When "den.sd" is set to "lnorm" then 'par.sd[,1]' is the vector of log(means) for the density and 'par.sd[,2]' is the vector of log(standard deviations) for the distributions. If there is only one regime fitted to the tree, then 'par.sd' is a vector with length 2.}

\item{unif.corr}{whether the correlation structure of the prior distribution on the Sigma matrix is flat. This sets an uniformative prior (as uninformative as possible) to the evolutionary correlations among the traits.}

\item{Sigma}{the scale matrix to be used as a parameter for the inverse-Wishart distribution responsible for the generation of the correlation matrices. Need to be a list of matrices with number of elements equal to the number of evolutionary rate regimes fitted to the phylogeny, in the case of a single regime, this needs to be a matrix (not a list). The scale matrix is somewhat analogous to the mean of a normal distribution. Thus, if 'Sigma' show strong positive correlation among traits, then the distribution generated by the prior will vary around positive correlations and, as a result, be significantly informative. This parameter will be ignored if 'unif.corr' is set to TRUE.}

\item{nu}{the degrees of freedom parameter for the inverse-Wishart distribution. Need to be a numeric vector with length equal to the number of evolutionary rate matrix regimes fitted to tree. Larger values of 'nu' will make the prior distribution closer around 'Sigma', where small values will increase the variance. This parameter is analogous to the variance parameter of a normal distribution, however, it has an inverted relationship.}
}
\value{
List of density functions that calculated the log-likelihood given parameter values, the log-likelihood of the prior is equal to the sum of the prior for the root value (first element), the prior for the correlation matrix (second element), and the prior for the vector of standard deviations (third element). The fourth element of the list is a list with the parameters used to specify the prior.
}
\description{
Generates prior densities for the MCMC sampler. This function is integrated within the 'ratematrixMCMC' function that runs the MCMC chain. However, this implementation allows for more control over the prior distribution for the analysis. The prior functions produced here can be easily passed to the 'ratematrixMCMC' function.
}
\details{
Independent priors are defined for the phylogenetic mean, the vector of standard deviations and the structure of correlation. Thus, allowing for a wide range of configurations. Priors for the phylogenetic mean and the standard deviations can be uniform or normal (lognormal in the case of the standard deviations). The prior on the matrix of correlations is distributed as an inverse-Wishart and can be set to a marginally uniform prior or to be centered around some custom variance-covariace matrix. \cr
\cr
The prior for the model has two elements, one is the vector of phylogenetic means and the other is the evolutionary rate matrix regimes fitted to different branches of the tree. The vector of root values can be distributed as any continuous distribution, in this implementation the two options are the uniform and the normal distribution. On the other hand, the prior distribution for the rate matrix need to be more elaborated. Here we divide the variance-covariance matrix into two elements, the correlation matrix and the vector of standard deviations. Standard deviations can be modelled as any continuous and positive distribution, here we use a uniform or a log-normal distribution. The correlation matrix need to be derived from a distribution of covariance matrices known as the inverse-Wishart. The inverse-Wishart is controlled by two parameters; the scale matrix (Sigma) and the degrees of freedom (nu). Any variance-covariance matrix can be used as the scale matrix. To set a marginally uniform prior for the correlation structure of the evolutionary rate matrices sampled for the model one need to set 'Sigma' as an identity matrix and 'nu' as the dimension of the matrix +1. This is performed automatically by the function when the option 'unif.corr' is set to TRUE.
}
\examples{
\donttest{
## Set the prior, take a sample from it and compute the log-likelihood.
par.mu <- rbind( c(-10, 10), c(-10, 10) )
par.sd <- rbind( c(0, 10), c(0, 10) )
prior <- makePrior(r=2, p=2, par.mu=par.mu, par.sd=par.sd)
sample.prior <- samplePrior(n=1, prior=prior)
## Reconstruct the variance-covariance matrix from the correlation matrix and the variances.
## Note that the model has two evolutionary rate matrix regimes. So we need a list.
R1 <- diag(sample.prior$sd[[1]]^2) \%*\% sample.prior$matrix[[1]] \%*\% diag(sample.prior$sd[[1]]^2)
R2 <- diag(sample.prior$sd[[2]]^2) \%*\% sample.prior$matrix[[2]] \%*\% diag(sample.prior$sd[[2]]^2)
R <- list( R1, R2 )

## Load the data
data( centrarchidae )
## Compute the log-likelihood.
likelihoodFunction(data=centrarchidae$data, phy=centrarchidae$phy.map, root=sample.prior$mu, R=R)

## Start a MCMC chain.
handle <- ratematrixMCMC(data=centrarchidae$data, phy=centrarchidae$phy.map, gen=1000, prior=prior)
posterior <- readMCMC(handle, burn=0.25, thin=1)
plotRatematrix(posterior)
}
}
\author{
Daniel S. Caetano and Luke J. Harmon
}
