% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePropMeanForMult.R
\name{makePropMeanForMult}
\alias{makePropMeanForMult}
\title{Proposal and accept/reject for the phylogenetic mean.}
\usage{
makePropMeanForMult(cache.data, cache.chain, prior, v, w_sd, w_mu, iter, count,
  traitwise = FALSE, use_corr = TRUE)
}
\arguments{
\item{cache.data}{list. The cache with the data.}

\item{cache.chain}{list. The cache with the MCMC chain.}

\item{prior}{list. A list with the prior functions.}

\item{v}{numeric. Degrees of freedom for the inverse-wishart distribution.}

\item{w_sd}{For standard use for the proposal of rate matrices. No use here.}

\item{w_mu}{Width of the proposal for the phylogenetic mean.}

\item{iter}{numeric. The state of the MCMC chain. This is used to access elements in the MCMC caches.}

\item{count}{numeric. Keep the count of the chain to record accepted and rejected steps.}

\item{traitwise}{Whether the proposal for the phylogenetic mean is made for all the traits or trait-by-trait.}

\item{use_corr}{Whether the proposal for the root value will use the correlation of the tip data to sample proposals following the major axis of variation observed in the data. When this is set to TRUE the update will use a multivariate normal distribution and, thus, will always sample the values for the two traits at once. The value for "traitwise" will be ignored, if "use_corr" is set to TRUE.}

\item{w}{numeric. The sliding window width parameter.}
}
\value{
Return a modified 'cache.chain'.
}
\description{
Make the proposal and acceptance steps for the phylogenetic mean.
}
\details{
Internal function to be used in the MCMC. This functions uses a simple sliding window strategy to perform the updates. This version will make independent updates for each of the phylogenetic means. This might improve the mixing.
}

