% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergePosterior.R
\name{mergePosterior}
\alias{mergePosterior}
\title{Merge posteriors}
\usage{
mergePosterior(..., p)
}
\arguments{
\item{...}{Any number of posterior distributions as produced by the function 'read.multi.R.iwish' or 'read.single.R.iwish'}

\item{p}{Number of R matrices (regimes) fitted to the tree.}
}
\value{
A merged posterior distribution in the same format.
}
\description{
Merge two or more posterior distributions
}
\details{
Merge two or more posterior distributions based in the same dataset. If you have multiple MCMC chains for the same dataset you can merge them.
}

