% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseResultsMCMC.R
\name{parseResultsMCMC}
\alias{parseResultsMCMC}
\title{Parse results of simulations.}
\usage{
parseResultsMCMC(out.one, out.two, chain.one, chain.two, flag = NULL)
}
\arguments{
\item{out.one}{output of the MCMC}

\item{out.two}{output of the MCMC}

\item{chain.one}{The MCMC chain one.}

\item{chain.two}{The MCMC chain two.}

\item{flag}{Name for the run.}
}
\value{
Plots and etc.
}
\description{
Function to read and parse the results from the simulations.
}
\details{
Function to read and parse the results from the simulations. At this point the function will only work
  when there is two chains. This need to be extended and fixed for it to run with one or two chains.
}

