% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrior.R
\name{plotPrior}
\alias{plotPrior}
\title{Plot the prior distribution used in the MCMC analysis.}
\usage{
plotPrior(out, n = 1000, color = "black", ...)
}
\arguments{
\item{out}{the output object from the 'ratematrixMCMC' function.}

\item{n}{number of samples from the prior to be plotted (default is 1000).}

\item{color}{color for the plot.}

\item{...}{parameters to be passed to the function 'plotRatematrix'. See help for 'plotRatematrix' for list of possible parameters.}
}
\value{
A plot similar to 'plotRatematrix'.
}
\description{
Function uses the output object from the MCMC chain performed by 'ratematrixMCMC' to plot samples from the prior distribution.
}
\author{
daniel
}

