% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRatematrix.R
\name{plotRatematrix}
\alias{plotRatematrix}
\title{Plot posterior distribution of rate matrices.}
\usage{
plotRatematrix(chain, p, colors, alphaOff = 1, alphaDiag = 1,
  alphaEll = 1, ell.wd = 0.5, point.matrix = NULL, point.color = NULL,
  point.wd = 0.5, leg = NULL, l.cex = 0.7, hpd = 100,
  show.zero = FALSE, set.xlim = NULL, n.lines = 50)
}
\arguments{
\item{chain}{The MCMC chain loaded from the files.}

\item{p}{A vector with the regimes to be plotted.}

\item{colors}{A vector with colors. Same length as p.}

\item{alphaOff}{Transparency of the off-diagonals plots. Numeric between 0 and 1.}

\item{alphaDiag}{Transparency of the diagonals plots. Numeric between 0 and 1.}

\item{alphaEll}{Transparency of the ellipses lines. Numeric between 0 and 1.}

\item{ell.wd}{The width of the ellipses lines.}

\item{point.matrix}{Optional. A list with matrices of length equal to p. This will be plotted as lines in the grid.plot.}

\item{point.color}{Optional. A vector of colors with length equal to p. This is a alternative color vector to the 'colors' argument to be used with the point matrices. If not provided then the colors of the point matrices will be equal to 'colors'.}

\item{point.wd}{Optional. The width of the lines for the vertical and ellipses of the point matrices.}

\item{leg}{string. Legend for the traits. Vector need to have length equal to the dimension of the R matrix.}

\item{l.cex}{numeric. 'cex' parameter for 'leg'. See 'help(par)' for more information on 'cex'.}

\item{hpd}{numeric. Set the proportion of the highest posterior density (HPD) to be highlighted in the plot. If set .95 the distributions and the ellipses within the 95\% HPD will be highlighted in the plot. Default value does not show highlighted region.}

\item{show.zero}{logical. Whether to plot a 'blue' vertical line at 0.}

\item{set.xlim}{numeric. Two elements vector to set the 'xlim' [see 'help(hist)'] for the density plots manually. If 'NULL', the default value, then the limits are calculated from the data.}

\item{n.lines}{numeric. Number of lines to be displayed in the ellipsed plots.}
}
\value{
Plot a grid of charts.
}
\description{
Make a grid plot with the posterior distribution of the evolutionary rate matrix (R). Function works with one or two matrices fitted to the same tree and will superpose the results.
}
\details{
The upper-tri plots are the posterior density for the off-diagonal elements of the
    R evolutionary matrix. The diagonal plots are the posterior elements of the evolutionary
    rates for each of the traits. The lower-tri plots are ellipse plots for the rates of correlated
    evolution.\cr
\cr
Two posterior distributions of matrices can be plotted in the same grid if a distribution for
    'mat2' is also provided (optional). Otherwise only 'mat1' is plotted.\cr
\cr
If a second distribution of matrices is provided ('mat2') the function will check whether the number
     of matrices in each list is the same. If not, the first elements of the lengthier list will be pruned
     to match the length of the shorter list.\cr
\cr
Function use 'par()' to set graphical parameters. It may take some time to create all the plots if
     the number of traits is large. Thinning the posterior distribution of matrices before plotting is
     recommended.
}

