% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postPredicSims.R
\name{postPredicSims}
\alias{postPredicSims}
\title{Make posterior predictive simulations.}
\usage{
postPredicSims(out, mcmc.chain, mc.cores = NULL)
}
\arguments{
\item{out}{The 'out' file from the MCMC function.}

\item{mcmc.chain}{The posterior chain generated by 'read.single.R.iwish' or 'read.multi.R.iwish'.}

\item{mc.cores}{The number of cores to be used for parallel simulations. Optional argument. Default is 'NULL' and uses a single computer processor. Will import 'parallel' if provided with a number of cores.}
}
\value{
A list of matrices. Each element is a simulation of datasets. Using the same model and parameter estimates from the joint posterior distribution.
}
\description{
Perform posterior predictive simulations.
}
\details{
This function performs posterior predictive simulations. The simulations can be used to check if the model and parameter estimates can generate data that is close to your data. This is a measure of model adequacy. We expect the most adequate model to generate data is closer to the observed data than the alternative models. After simulating a large number of datasets you can use a series of summary statistics to check how close the simulated data is to the observed data. Check the examples.
}
\seealso{
\code{read.single.R.iwish} or \code{read.multi.R.iwish}
}

