% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratematrix-package.R
\docType{package}
\name{ratematrix}
\alias{ratematrix}
\alias{ratematrix-package}
\title{ratematrix.}
\description{
Package to estimate the evolutionary rate matrix for multiple regimes fitted to the phylogenetic tree using Bayesian Markov-chain Monte Carlo.
}
