% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratematrixPolytopeMCMC.R
\name{ratematrixPolytopeMCMC}
\alias{ratematrixPolytopeMCMC}
\title{Joint estimate the multidimentional phenotypic space for ancestrals and the evolutionary rate matrix}
\usage{
ratematrixPolytopeMCMC(data, phy, sample_internal = FALSE,
  save_start_anc = TRUE, rate_variation = FALSE, ncat = 10,
  prior = "uniform_scaled", start = "prior_sample", gen, burn = 0.25,
  thin = 100, v = 50, w_sd = 0.5, w_gamma = 0.5,
  prop_par = c(0.2, 0.2, 0.2, 0.2, 0.2), n_tips_move = 1, dir = NULL,
  outname = "ratematrixPolyMCMC", IDlen = 5, save.handle = TRUE)
}
\arguments{
\item{data}{a named list with multiple trait values per species. Names of list elements need to match tip labels. list of matrices with observations as rows and trait values as columns. See "Details" for more information.}

\item{phy}{a single phylogeny of the class "simmap" with the mapped regimes for two or more R regimes OR a phylogeny of the class "phylo" for a single regime. The number of evolutionary rate matrices fitted to the phylogeny is equal to the number of regimes in 'phy'. Regime names will also be used.}

\item{sample_internal}{whether to sample the internal nodes of the tree (i.e., ancestral states) using Gibbs sampler together with the estimates of the other parameter of the model. NOTE: Preliminary results showed that including this parameter can decrease the mixing of the MCMC chain.}

\item{save_start_anc}{whether to save the starting state for the internal nodes. This can be used as the starting state for search replicates using the same data. See 'start' argument for more information. This is ignored if 'sample_internal = FALSE'.}

\item{rate_variation}{whether to allow for rate heterogeity on the model. Rate variation is modelled using a discrete-Gamma distribution (Yang, 1994). This is similar to the +G models of molecular evolution.}

\item{ncat}{number of categories of the discrete-Gamma distribution for variation of rates. Large numbers allow for more smooth rate variation but make convergence of the model slower.}

\item{prior}{the prior densities for the MCMC. Must be one of "uniform", "uniform_scaled" (the default, see 'Details'), or the output of the "makePrior" function. See more information on 'makePrior' and in the examples below.}

\item{start}{the starting state for the MCMC chain. Must be "prior_sample" (the default) or a list in the same format as the prior sample generated from the "samplePrior" function. The function will also search for a "$anc" element on the list to use as the starting state for the internal nodes (if 'sample_internal = TRUE' ).}

\item{gen}{number of generations for the chain.}

\item{burn}{percentage to discard as burn in. Set 'burn' in to 0.0 to start writing from the first generation.}

\item{thin}{number of generations to be skipped as thinning. Set 'thin' to 1 to write every generation to the file.}

\item{v}{value for the degrees of freedom parameter of the inverse-Wishart proposal distribution for the correlation matrix. Smaller values provide larger steps and larger values provide smaller steps. (Yes, it is counterintuitive.) This needs to be a single value applied to all regimes or a vector with the same length as the number of regimes.}

\item{w_sd}{the multiplying factor for the multiplier proposal on the vector of standard deviations. This can be a single value to be used for the sd of all traits for all regimes or a matrix with number of columns equal to the number of regimes and number of rows equal to the number of traits. If a matrix, then each element will be used to control the correspondent width of the standard deviation.}

\item{w_gamma}{the multiplying factor for the multiplier proposal on the rate parameter for the Gamma distributed rate variation across the branches of the tree.}

\item{prop_par}{a numeric vector of length 5 with the proposal frequencies for the parameters in this order: tip states, variance (rates), correlation, Gamma parameter (beta), and samples from the Gamma rate distribution. Default value is 'c(0.2, 0.2, 0.2, 0.2, 0.2)'.}

\item{n_tips_move}{the number of nodes (among tip nodes and internal nodes) that are going to be updated during each step of the MCMC. Min of 1 and max should be no more than the number of tips in the tree.}

\item{dir}{path of the directory to write the files. Has no default value (due to RCran policy). The path can be provided both as relative or absolute. It should accept Linux, Mac and Windows path formats.}

\item{outname}{name for the MCMC chain (default is 'ratematrixMCMC'). Name will be used in all the files alongside a unique ID of numbers with length of 'IDlen'.}

\item{IDlen}{length of digits of the numeric identifier used to name output files (default is 5).}

\item{save.handle}{whether the handle for the MCMC should be saved to the directory in addition to the output files.}
}
\value{
Function returns the 'handle' object and writes the posterior distribution and log as files in the directory (see 'dir'). The handle is a list with the details of the MCMC chain. It is composed by: *k* the number of traits; *p* the number of R regimes fitted to the tree; *ID* the unique identifier of the run; *dir* the directory where the posterior and log files were saved; *outname* the name for the chain; *trait.names* a vector with the label for the traits; *regime.names* a vector with the label for the rate regimes; *data* the data used in the analysis; *phy* a single phylogeny or the list of phylogenies; *prior* a list with the prior functions; *start* a list with the starting parameters for the chain; *gen* the number of generations for the chain; *mcmc.par* a list with the tunning parameters for the MCMC.
}
\description{
Function runs a MCMC chain to estimate the multivariate phenotypic space occupied by the ancestrals given the data observed at the tips of the phylogeny and the model of trait macroevolution. The trait evolution model incorporates separated rates of evolution for each trait as well as the evolutionary correlation among the traits. This method is an extension of the model described in Caetano and Harmon (2018). The multivariate phenotypic space is treated as a n-dimensional polytope. We use data augmentation to estimate the ancestral states following previous work by Lemey et al. (2010) and Bouckaert et al. (2012).
}
\details{
The input data is a named list with a matrix of observations for each of the species on the tree. For each matrix, the rows are individual observations and the columns are each of the traits. The number of rows between matrices can vary. At least two observations are necessary for each of the species. The number of columns need to be the same and in the same order among matrices. The names of the list elements need to match the names of the species in the tree. Names for the columns of each matrix are optional.

This function does not allow for a list of phylogenies as the input. Since the aim is to estimate the phenotypic space occupied by ancestors, it is unclear how to summarize information across distinct topologies.

Please check 'ratematrixMCMC' for more details. This function follows the same algorithm in 'ratematrixMCMC' except for the data augmentation approach used to incorportate the phenotypic space for the traits at the tips and nodes of the tree. Also note that this function does not estimate the root value for the tree. Here we use the restricted likelihood for the multivariate Brownian motion model.
}
\examples{
\donttest{
## Need to add examples.
}
}
\references{
Revell, L. J., and L. J. Harmon. 2008. Testing quantitative genetic hypotheses about the evolutionary rate matrix for continuous characters. Evolutionary Ecology Research 10:311.

Revell, L. J., and D. C. Collar. 2009. Phylogenetic Analysis of the Evolutionary Correlation Using Likelihood. Evolution 63:1090–1100.

Lemey, P., A. Rambaut, J. J. Welch, and M. A. Suchard. 2010. Phylogeography Takes a Relaxed Random Walk in Continuous Space and Time. Mol Biol Evol 27:1877–1885.

Bouckaert, R., P. Lemey, M. Dunn, S. J. Greenhill, A. V. Alekseyenko, A. J. Drummond, R. D. Gray, M. A. Suchard, and Q. D. Atkinson. 2012. Mapping the Origins and Expansion of the Indo-European Language Family. Science 337:957–960.

Caetano, D. S., and L. J. Harmon. 2017. ratematrix: An R package for studying evolutionary integration among several traits on phylogenetic trees. Methods in Ecology and Evolution 8:1920–1927.

Caetano, D. S., and L. J. Harmon. 2018. Estimating Correlated Rates of Trait Evolution with Uncertainty. Systematic Biology, doi: 10.1093/sysbio/syy067.

Yang, Z. 1994. Maximum likelihood phylogenetic estimation from DNA sequences with variable rates over sites: Approximatemethods. Journal of Molecular Evolution 39:306-314.
}
\author{
Daniel S. Caetano and Luke J. Harmon
}
