% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R.matrix.step.zhang.R
\name{sigma.step.zhang}
\alias{sigma.step.zhang}
\title{Sigma step using Zhang strategy}
\usage{
\method{sigma}{step.zhang}(cache.data, cache.chain, prior, w, v, iter, count)
}
\arguments{
\item{cache.data}{}

\item{cache.chain}{}

\item{prior}{}

\item{w}{}

\item{iter}{}

\item{count}{}
}
\value{
The chain cache.
}
\description{
Description!
}
\details{
Details!
}
\author{
daniel
}

