% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleR.loglik.R
\name{singleR.loglik}
\alias{singleR.loglik}
\title{Log likelihood for one evolutionary rate matrix fitted to the data and tree.}
\usage{
singleR.loglik(X, root, R, C, D, n, r, method = c("rpf", "sparse"))
}
\arguments{
\item{X}{matrix. The matrix of trait data.}

\item{root}{numeric. The phylogenetic mean.}

\item{R}{matrix. The covariance matrix fitted to the tree and data.}

\item{C}{matrix. The phylogenetic covariance matrix. This is a fixed term.}

\item{D}{matrix. The design matrix. This is a fixed term, dependent on the phylogeny.}

\item{n}{numeric. Number of tips in the phylogeny.}

\item{r}{numeric. Number of traits, the dimension of the R matrix.}

\item{method}{string. Method to be used to calculate the likelihood. Right now only the option 'rpf' is working. More options will be implemented in the future.}
}
\value{
numeric. The log likelhood.
}
\description{
Log likelihood function for one evolutionary rate matrices.
}
\details{
Internal function to be used by the MCMC.
}
\references{
Revell, L. J., and D. C. Collar. 2009. Phylogenetic Analysis of the Evolutionary Correlation Using Likelihood. Evolution 63:1090–1100.
}

