% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testRatematrix.R
\name{testRatematrix}
\alias{testRatematrix}
\title{Test for difference between evolutionary rate matrix estimates}
\usage{
testRatematrix(chain, par = c("all", "correlation", "rates"),
  median.test = FALSE, regimes = NULL, plot = TRUE)
}
\arguments{
\item{chain}{the posterior distribution of parameter estimates as output of the function 'readMCMC'.}

\item{par}{sets which of the attributes of the rate matrices that are checked by the test. One of 'all', 'correlation', or 'rates' (default is 'all'). Chose 'all' to test the median overlap among all cells of the covariance matrix. Chose 'rates' to check for the median overlap among the evolutionary rates of each trait (the variance vector). Chose 'correlation' to test for the median overlap among the correlations (the distribution of correlation matrices).}

\item{median.test}{whether to return a median of the summary statistics across all elements of the evolutionary rate matrices. The default is FALSE.}

\item{regimes}{a numeric vector or character vector. This is the set of regimes to be compared. If numeric, then the regimes are in the same order as in the 'chain' argument. If character, than names need to match the names of the rate regimes fitted to the phylogenetic tree.}

\item{plot}{whether to plot the results of the summary statistics test applied to every element of the matrix.}
}
\value{
Return a matrix or a list with the value of the test statistics. Values below 0.05 support a difference between the posterior distribution of R matrices. When computing the summary statistics for each element of the evolutionary rate matrix (i.e., median.test=FALSE) the function returns a list object with element-wise results for each regime combination and also plots the results in a heatmap format.
}
\description{
Function uses summary statistics to test for differences between the posterior distribution of parameter estimates for the evolutionary rate matrix regimes.
}
\details{
This functions performs a test to check whether the posterior distribution of the fitted matrices are different. It returns the proportion of overlap between regimes. When this proportion is less than 0.05 this means that the posterior distribution of the elements of the evolutionary rate matrices does not overlap more than 5\%. This test is not a p value! This is not an estimate of a probability of deviance from a given null distribution. It assumes that when the posterior distribution of two or more paramaters do not overlap, then there is a relevant difference between the parameters. \cr
The test can be performed using the median overlap of the posterior distribution across all elements of the ratematrix or by contrasting each element separatelly. Checking each element independently provides more information. Using the median overlap will result in a single value returned, but it can be insensitive to important changes in the evolutionary rate matrices between regimes. \cr
\cr
When a posterior distribution with more than two rate regimes is fitted to the data, the function performs tests for all pairwise combinations.
}
\examples{
\donttest{
## Load data with posterior distribution.
data(anolesPost)
testRatematrix(chain=anolesPost$chain1, par="all")
testRatematrix(chain=anolesPost$chain1, par="rates")
testRatematrix(chain=anolesPost$chain1, par="correlation")
## Compare with results when using the median of the summary statistics:
testRatematrix(chain=anolesPost$chain1, par="all", median.test=TRUE)
testRatematrix(chain=anolesPost$chain1, par="rates", median.test=TRUE)
testRatematrix(chain=anolesPost$chain1, par="correlation", median.test=TRUE)

}
}
\author{
Daniel S. Caetano and Luke J. Harmon
}
